1. 新增相位瞬移 2. Ifrit AI調整 3. 新增會碎裂的平台 4. 鏡頭晃動調整 5. 現在攻擊會造成鏡頭晃動了

+++ 掠食者/Assets/Script/Character/Skills/Controller/SkillController.cs
@@ -36,7 +36,7 @@ public class SkillController : MonoBehaviour
                 {
                     return false;
                 }
-                character.CurrentHealth -= skill.cost;
+                character.CurrentHealth -= (int)skill.cost;
                 break;
             case CostType.Mana:
                 if (character.CurrentMana < skill.cost)
